oUF_NeavRaid Config File
The ingame config is now made with lua templates instead of xml. This should make it easier to understand and to add more features in the future.

This also makes localization easier. The old ui was using a mix of global string and hardcoded english. The new localization file will give us total control so everyone can have a properly translated ui. We will need people to submit translations for it shouldn't be that hard.